refactor: Rename `lib/paramCommand.js`
