+++ lib/param-command.js
