+++ lib/plugin.js
@@ -20,7 +20,7 @@ const { configureDeployProfile } = require('./deploy-profile');
 const { test } = require('./test');
 const { getDashboardUrl, dashboardHandler } = require('./dashboard');
 const setApiGatewayAccessLogFormat = require('./setApiGatewayAccessLogFormat');
-const paramCommand = require('./paramCommand');
+const paramCommand = require('./param-command');
 const outputCommand = require('./output-command');
 const isAuthenticated = require('./is-authenticated');
 const throwAuthError = require('./throwAuthError');
